🔧 Update ESLint rules for `if` statements
JavaScript allows the omission of curly braces
when a block contains only one statement.

However, it is considered by many to be best
practice to never omit curly braces around blocks,
even when they are optional, because it can lead to
bugs and reduces code clarity.

Omission is noticed mostly with `if (<statement>) continue;`
constructs, but curly braces should not be omitted.

Also if an `if` statement is the only statement in the
`else` block, it is often clearer to use an `else if` form.

Lastly, constructs like
```
if (x) {
  return y;
} else {
  return z;
}
```
should omit else branch and use `return` directly.